requirements.txt
streamlit
google-cloud-translate
google-cloud-core
google-auth
